fix data of type in entities

+++ SwimmingPoolServiceExample/src/main/java/com/inbadevs/swimmingpoolserviceusers/buisness/ManagerSchedule.java
@@ -32,10 +32,9 @@ public class ManagerSchedule {
     
     public void modifySchedule(Schedule schedule){
         this.schedule.update(schedule);
-    
     }
     
-    public void deleteSchedule(Integer id){
+    public void deleteSchedule(Long id){
         this.schedule.delete(id);
     
     }
